move files into folder (#667)
